refactor(web): move events domain from ducks/ to src/events
Pure file move: ducks/events/{queries,mutations,event.api.ts,
event.types.ts} relocate to src/events/ alongside the new Zustand
stores, dissolving the Redux-era ducks/ directory. Imports rewritten
from @web/ducks/events/* to @web/events/*; no logic changes.
