+++ packages/web/src/events/queries/event.query.normalize.ts
