+++ packages/web/src/events/queries/event.query.options.ts
