+++ packages/web/src/events/queries/event.view-model.ts
