+++ packages/web/src/events/queries/useDayEventsQuery.ts
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { useEffect, useMemo } from "react";
 import { useEventRepositorySource } from "@web/common/repositories/event/event.repository.source.store";
 import { handleError } from "@web/common/utils/event/event.util";
-import { dayEventsQueryOptions } from "@web/ducks/events/queries/event.query.options";
+import { dayEventsQueryOptions } from "@web/events/queries/event.query.options";
 import { deriveCalendarEventViewModel } from "./event.view-model";
 
 type DayEventsQueryArgs = {
