+++ packages/web/src/events/queries/usePrefetchAdjacentEvents.ts
