+++ packages/web/src/sse/hooks/useGcalSSE.ts
@@ -3,7 +3,7 @@ import { handleGoogleRevoked } from "@web/auth/google/util/google.auth.util";
 import { userMetadataActions } from "@web/auth/state/user-metadata.store";
 import { queryClient } from "@web/common/query/query-client";
 import { showErrorToast } from "@web/common/utils/toast/error-toast.util";
-import { eventQueryKeys } from "@web/ducks/events/queries/event.query.keys";
+import { eventQueryKeys } from "@web/events/queries/event.query.keys";
 import { sseEmitter } from "../client/sse.client";
 import { createUseGcalSSE } from "./useGcalSSE.factory";
